+++ apps/cowswap-frontend/src/modules/affiliate/pure/TraderIneligible.tsx
@@ -1,0 +1,16 @@
+import { ReactNode } from 'react'
+
+import { Trans } from '@lingui/react/macro'
+
+import { HowItWorks } from './HowItWorks'
+
+export function TraderIneligible(): ReactNode {
+  return (
+    <>
+      <Trans>
+        This wallet has already traded on CoW Swap. <br /> Referral rewards are for new wallets only.
+      </Trans>{' '}
+      <HowItWorks />
+    </>
+  )
+}