+++ libs/analytics/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [2.3.2](https://github.com/cowprotocol/cowswap/compare/analytics-v2.3.1...analytics-v2.3.2) (2026-02-25)
+
+
+### 🔧 Miscellaneous
+
+* main -&gt; dev ([9c6611e](https://github.com/cowprotocol/cowswap/commit/9c6611e90a4b23e23f6aaadc513de3c6d7f3b562))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/common-const bumped to 2.3.2
+    * @cowprotocol/common-hooks bumped to 2.3.2
+    * @cowprotocol/common-utils bumped to 2.3.2
+    * @cowprotocol/events bumped to 3.3.2
+
 ## [2.3.1](https://github.com/cowprotocol/cowswap/compare/analytics-v2.3.0...analytics-v2.3.1) (2026-02-24)